requirements.txt
python-telegram-bot  
python-dotenv  
solana
solders

